+++ src/types/inquirer-autocomplete-prompt.d.ts
@@ -1,0 +1,4 @@
+declare module 'inquirer-autocomplete-prompt' {
+    const autocompletePrompt: unknown;
+    export default autocompletePrompt;
+}